수익률을 계산한다. 총 수익률에 손해, 이익을 표시한다.

+++ src/main/kotlin/lotto/service/LottoShop.kt
@@ -8,7 +8,7 @@ class LottoShop {
     companion object {
         private const val LOTTO_PRICE = 1000
 
-        fun purchase(purchaseMoney: Int): Lottos {
+        fun purchase(purchaseMoney: Long): Lottos {
             val purchaseCount = purchaseCount(purchaseMoney)
 
             return Lottos(
@@ -18,7 +18,7 @@ class LottoShop {
             )
         }
 
-        private fun purchaseCount(purchaseMoney: Int): Int {
+        private fun purchaseCount(purchaseMoney: Long): Long {
             return purchaseMoney / LOTTO_PRICE
         }
     }
