+++ src/main/kotlin/lotto/view/InputView.kt
@@ -4,9 +4,9 @@ import lotto.domain.Lotto
 
 class InputView {
     companion object {
-        fun purchaseMoney(): Int {
+        fun purchaseMoney(): Long {
             println("구입금액을 입력해 주세요.")
-            return readln().toInt()
+            return readln().toLong()
         }
 
         fun winningLotto(): Lotto {
